Added user type checks on endpoints.

+++ InterCareBackend/Controllers/ManagerController.cs
@@ -1,19 +1,30 @@
 ﻿using InterCareBackend.Daos.Implementations;
+using InterCareBackend.Helpers;
 using InterCareBackend.Models;
 using Microsoft.AspNetCore.Mvc;
+using System.Collections.Generic;
 
 namespace InterCareBackend.Controllers
 {
-    public class ManagerController
+    public class ManagerController : ControllerBase
     {
 
         LocationManagerDao managerDao = new LocationManagerDao();
-
+        AuthHelper auth = new AuthHelper();
 
         [HttpPost("/api/createLocationManager")]
-        public void create()
+        public string create()
         {
-            managerDao.createManager("Jonar17@student.sdu.dk", "Pass", "Jonas Støve Rasmussen", "0", "Manager");
+            var header = Request.Headers["Authorization"].ToString().Substring("Bearer ".Length).Trim();
+            IDictionary<string, object> token = this.auth.decodeJWT(header);
+            if (token["type"].ToString() == Globals.GlobalLocationManager)
+            {
+                managerDao.createManager("Jonar17@student.sdu.dk", "Pass", "Jonas Støve Rasmussen", "0", "Manager");
+                return "Manager has been created!";
+            } else
+            {
+                return "Unable to create manager.";
+            }
         }
 
 
@@ -21,14 +32,31 @@ namespace InterCareBackend.Controllers
 
         public LocationManager getManagerByEmail()
         {
-            return managerDao.getLocationManagerByEmail("Jonar17@student.sdu.dk");
+            var header = Request.Headers["Authorization"].ToString().Substring("Bearer ".Length).Trim();
+            IDictionary<string, object> token = this.auth.decodeJWT(header);
+            if (token["type"].ToString() == Globals.GlobalLocationManager)
+            {
+                return managerDao.getLocationManagerByEmail("Jonar17@student.sdu.dk");
+            } else
+            {
+                return null;
+            }
         }
 
 
         [HttpDelete("/api/deleteLocationManagerByEmail")]
-        public void delete()
+        public string delete()
         {
-            managerDao.deleteManager("Jonar17@student.sdu.dk");
+            var header = Request.Headers["Authorization"].ToString().Substring("Bearer ".Length).Trim();
+            IDictionary<string, object> token = this.auth.decodeJWT(header);
+            if (token["type"].ToString() == Globals.GlobalOrganizationAdmin || token["type"].ToString() == Globals.GlobalInterCareAdmin)
+            {
+                managerDao.deleteManager("Jonar17@student.sdu.dk");
+                return "Manager has been deleted";
+            } else
+            {
+                return "Could not delete manager!";
+            }
         }
 
     }
